Add option to skip redownloading
Previously, you could only specify whether you wanted to download
images or not. If a block was flagged to download images, then they
would always be downloaded, and then later anki would save the image
or not depending on whether it existed. Now the option is further up
the stack, so images will only be redownloaded if specified.